// web/src/api/models/RepositoryGetMembersByHouseholdIDRow.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RepositoryGetMembersByHouseholdIDRow
 */
export interface RepositoryGetMembersByHouseholdIDRow {
    /**
     * 
     * @type {string}
     * @memberof RepositoryGetMembersByHouseholdIDRow
     */
    householdId?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryGetMembersByHouseholdIDRow
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryGetMembersByHouseholdIDRow
     */
    joinedAt?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryGetMembersByHouseholdIDRow
     */
    role?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryGetMembersByHouseholdIDRow
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryGetMembersByHouseholdIDRow
     */
    username?: string;
}

/**
 * Check if a given object implements the RepositoryGetMembersByHouseholdIDRow interface.
 */
export function instanceOfRepositoryGetMembersByHouseholdIDRow(value: object): value is RepositoryGetMembersByHouseholdIDRow {
    return true;
}

export function RepositoryGetMembersByHouseholdIDRowFromJSON(json: any): RepositoryGetMembersByHouseholdIDRow {
    return RepositoryGetMembersByHouseholdIDRowFromJSONTyped(json, false);
}

export function RepositoryGetMembersByHouseholdIDRowFromJSONTyped(json: any, ignoreDiscriminator: boolean): RepositoryGetMembersByHouseholdIDRow {
    if (json == null) {
        return json;
    }
    return {
        
        'householdId': json['household_id'] == null ? undefined : json['household_id'],
        'id': json['id'] == null ? undefined : json['id'],
        'joinedAt': json['joined_at'] == null ? undefined : json['joined_at'],
        'role': json['role'] == null ? undefined : json['role'],
        'userId': json['user_id'] == null ? undefined : json['user_id'],
        'username': json['username'] == null ? undefined : json['username'],
    };
}

export function RepositoryGetMembersByHouseholdIDRowToJSON(json: any): RepositoryGetMembersByHouseholdIDRow {
    return RepositoryGetMembersByHouseholdIDRowToJSONTyped(json, false);
}

export function RepositoryGetMembersByHouseholdIDRowToJSONTyped(value?: RepositoryGetMembersByHouseholdIDRow | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'household_id': value['householdId'],
        'id': value['id'],
        'joined_at': value['joinedAt'],
        'role': value['role'],
        'user_id': value['userId'],
        'username': value['username'],
    };
}