// web/src/api/models/RepositoryHousehold.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RepositoryHousehold
 */
export interface RepositoryHousehold {
    /**
     * 
     * @type {string}
     * @memberof RepositoryHousehold
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryHousehold
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryHousehold
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryHousehold
     */
    ownerId?: string;
}

/**
 * Check if a given object implements the RepositoryHousehold interface.
 */
export function instanceOfRepositoryHousehold(value: object): value is RepositoryHousehold {
    return true;
}

export function RepositoryHouseholdFromJSON(json: any): RepositoryHousehold {
    return RepositoryHouseholdFromJSONTyped(json, false);
}

export function RepositoryHouseholdFromJSONTyped(json: any, ignoreDiscriminator: boolean): RepositoryHousehold {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'ownerId': json['owner_id'] == null ? undefined : json['owner_id'],
    };
}

export function RepositoryHouseholdToJSON(json: any): RepositoryHousehold {
    return RepositoryHouseholdToJSONTyped(json, false);
}

export function RepositoryHouseholdToJSONTyped(value?: RepositoryHousehold | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'],
        'id': value['id'],
        'name': value['name'],
        'owner_id': value['ownerId'],
    };
}