// web/src/api/models/RepositorySearchUsersByUsernameRow.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RepositorySearchUsersByUsernameRow
 */
export interface RepositorySearchUsersByUsernameRow {
    /**
     * 
     * @type {string}
     * @memberof RepositorySearchUsersByUsernameRow
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositorySearchUsersByUsernameRow
     */
    username?: string;
}

/**
 * Check if a given object implements the RepositorySearchUsersByUsernameRow interface.
 */
export function instanceOfRepositorySearchUsersByUsernameRow(value: object): value is RepositorySearchUsersByUsernameRow {
    return true;
}

export function RepositorySearchUsersByUsernameRowFromJSON(json: any): RepositorySearchUsersByUsernameRow {
    return RepositorySearchUsersByUsernameRowFromJSONTyped(json, false);
}

export function RepositorySearchUsersByUsernameRowFromJSONTyped(json: any, ignoreDiscriminator: boolean): RepositorySearchUsersByUsernameRow {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'username': json['username'] == null ? undefined : json['username'],
    };
}

export function RepositorySearchUsersByUsernameRowToJSON(json: any): RepositorySearchUsersByUsernameRow {
    return RepositorySearchUsersByUsernameRowToJSONTyped(json, false);
}

export function RepositorySearchUsersByUsernameRowToJSONTyped(value?: RepositorySearchUsersByUsernameRow | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'username': value['username'],
    };
}